+++ src/rer/solrpush/tests/test_events.py
@@ -36,7 +36,7 @@ class TestSOLRPush(unittest.TestCase):
         setRoles(self.portal, TEST_USER_ID, ["Manager"])
         set_registry_record(
             "enabled_types",
-            ["Document", "File"],
+            [u"Document", u"File"],
             interface=IRerSolrpushSettings,
         )
 
@@ -81,7 +81,7 @@ class TestEvents(unittest.TestCase):
         setRoles(self.portal, TEST_USER_ID, ["Manager"])
         set_registry_record(
             "enabled_types",
-            ["Document", "File"],
+            [u"Document", u"File"],
             interface=IRerSolrpushSettings,
         )
         init_solr_push()